Factor out the library test
