+++ tutorials/advanced/library/lib/src/get_os_release.sh
